Add GRUB and GCC cross compiler support.
All linux based OS and most of the mordern operating systems follow the
Multiboot specification. This allows bootloaders like GRUB to boot from
multiple OSes present on the same drive. Hence for better portability
add a multiboot header at the start of the kernel which lets GRUB
identify us.
	Further, when a normal compiler such as GCC compiles C files, it
automatically assumes the presence of libc, the C library. This C
library contains important functions like printf(), but it is provided
by the OS. When we build our kernel binaries, we don't want GCC to link with
libc since it isn't present in our OS(yet). To fix this, shift to a GCC
cross compiler that does not assume the presence of libc. This change
also changes the assembly syntax from Intel to AT&T as nasm was used
earlier.