Add Rust kernel
Try to call the Rust kernel code after booting into long mode.
For some reason, this is not working as intended. We will be shifting
to a different bootloader after this commit due to deadlines but I intend
to come back and fix this :)

This commit adds code to call the _start() function which is the entry
point of our kernel. The problems encountered till now are:-

- Rust symbol tables simply doesn't record our function if it's
  named anything other than _start() ( Maybe has something to do with
  #![no_main] annotation

- Even when we use _start() and are able to link successfully, the code
  inside _start() simply doesn't execute. The assembly code after the
  function returns doesn't execute either. Maybe the symbols are not
  correctly defined and the call to _start() is jumping to an inccorect
  address.

TODO: Fix above problems